refactor: change authentiation request from User to Dto
