+++ src/main/java/com/karol/kindergartenmanagementsystem/dto/AuthenticationResponse.java
@@ -1,10 +1,11 @@
-package com.karol.kindergartenmanagementsystem.model;
+package com.karol.kindergartenmanagementsystem.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
+
 public class AuthenticationResponse {
     private String token;
 }
