+++ src/main/java/com/karol/kindergartenmanagementsystem/dto/SignInRequest.java
@@ -1,0 +1,15 @@
+package com.karol.kindergartenmanagementsystem.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignInRequest {
+    private String email;
+    private String password;
+}
